+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## (IN PROGRESS)
 
 * Add personal data disclosure form. Refs UIORGS-222.
+* Interface credentials not saving when create interface. Refs UIORGS-224
 * Contact information categories are not translated. Refs UIORGS-219.
 * Migrate Organization and Contact Forms to React-final-form. Refs UIORGS-200.